fix query param for acs proxy
