+++ controllers/common.go
@@ -3,14 +3,15 @@ package controllers
 import (
 	"encoding/json"
 	"errors"
-	"github.com/UHERO/rest-api/common"
-	"github.com/UHERO/rest-api/data"
-	"github.com/UHERO/rest-api/models"
-	"github.com/gorilla/mux"
 	"log"
 	"net/http"
 	"strconv"
 	"strings"
+
+	"github.com/UHERO/rest-api/common"
+	"github.com/UHERO/rest-api/data"
+	"github.com/UHERO/rest-api/models"
+	"github.com/gorilla/mux"
 )
 
 func CheckCache(c *data.CacheRepository) func(http.ResponseWriter, *http.Request, http.HandlerFunc) {
@@ -131,7 +132,7 @@ func getIdsList(w http.ResponseWriter, r *http.Request) (ids []int64, ok bool) {
 	ok = true
 	idsList, gotIds := mux.Vars(r)["ids_list"]
 	if !gotIds {
-		common.DisplayAppError(w, errors.New("Couldn't get id from request"),"Bad request.", 400)
+		common.DisplayAppError(w, errors.New("Couldn't get id from request"), "Bad request.", 400)
 		ok = false
 		return
 	}
@@ -139,7 +140,7 @@ func getIdsList(w http.ResponseWriter, r *http.Request) (ids []int64, ok bool) {
 	for _, idStr := range idStrArr {
 		id, err := strconv.ParseInt(idStr, 10, 64)
 		if err != nil {
-			common.DisplayAppError(w, err,"An unexpected error has occurred",500)
+			common.DisplayAppError(w, err, "An unexpected error has occurred", 500)
 			ok = false
 			return
 		}
@@ -286,6 +287,22 @@ func getIdGeoAndFreq(w http.ResponseWriter, r *http.Request) (id int64, geo stri
 		)
 		ok = false
 		return
+	}
+	return
+}
+
+func getAcsIdsList(w http.ResponseWriter, r *http.Request) (ids []string, ok bool) {
+	ok = true
+	idsList, gotIds := mux.Vars(r)["ids_list"]
+	if !gotIds {
+		common.DisplayAppError(w, errors.New("Couldn't get id from request"), "Bad request.", 400)
+		ok = false
+		return
+	}
+	idStrArr := strings.Split(idsList, ",")
+	for _, idStr := range idStrArr {
+		id := idStr
+		ids = append(ids, id)
 	}
 	return
 }
